changed js files to ts files
