+++ package.json
@@ -9,7 +9,7 @@
     "test:watch": "jest --watchAll --no-cache",
     "build": "tsc",
     "start": "node dist/index.js",
-    "start:watch": "tsc -w & nodemon dist/index.js"
+    "start:watch": "tsc -w & nodemon dist/server.js"
   },
   "author": "",
   "license": "ISC",
@@ -29,7 +29,11 @@
     "typeorm": "^0.2.29"
   },
   "devDependencies": {
+    "@shelf/jest-mongodb": "^1.2.3",
+    "@types/express": "^4.17.9",
     "@types/jest": "^26.0.19",
+    "@types/mongoose": "^5.10.3",
+    "@types/node": "^14.14.14",
     "@typescript-eslint/eslint-plugin": "^4.7.0",
     "@typescript-eslint/parser": "^4.7.0",
     "apollo-server-testing": "^2.19.0",
@@ -42,13 +46,12 @@
     "eslint-plugin-react": "^7.21.5",
     "eslint-plugin-react-hooks": "^4.2.0",
     "jest": "^26.6.0",
+    "mongodb-memory-server": "^6.9.2",
     "nodemon": "^2.0.6",
     "prettier": "^2.1.2",
+    "supertest": "^6.0.1",
     "ts-jest": "^26.4.4",
-    "typescript": "^4.0.5",
-    "@shelf/jest-mongodb": "^1.2.3",
-    "mongodb-memory-server": "^6.9.2",
-    "supertest": "^6.0.1"
+    "typescript": "^4.0.5"
   },
   "jest": {
     "testEnvironment": "node"
