+++ src/models/Tag.ts
@@ -1,3 +1,4 @@
+export {}
 const mongoose = require("mongoose");
 
 const TagSchema = new mongoose.Schema({
