+++ src/models/Topic.ts
@@ -1,3 +1,4 @@
+export {}
 const mongoose = require('mongoose');
 
 const TopicSchema = new mongoose.Schema({
