+++ src/models/User.ts
@@ -1,3 +1,4 @@
+export {}
 const mongoose = require("mongoose");
 const UserSchema = new mongoose.Schema({
   name: {
