+++ tsconfig.json
@@ -66,5 +66,5 @@
     "skipLibCheck": true /* Skip type checking of declaration files. */,
     "forceConsistentCasingInFileNames": true /* Disallow inconsistently-cased references to the same file. */
   },
-  "exclude": ["./src/**/*.spec.ts", "node_modules"]
+  "exclude": ["./src/**/*.spec.ts", "node_modules", "./src/z.*"]
 }
